Cleans models
