+++ clashstats/model/player.py
@@ -1,0 +1,12 @@
+from mongoengine import *
+
+
+class Player(DynamicDocument):
+    meta = {
+        'indexes': [
+            'clan.name',
+            'clan.tag',
+            'name',
+            'tag'
+        ]
+    }
