+++ clashstats/model/status.py
@@ -1,0 +1,9 @@
+from datetime import datetime
+
+from mongoengine import *
+
+
+class Status(Document):
+    last_updated = DateTimeField(default=datetime.now)
+    total_clans = IntField()
+    ratio_indexed = FloatField()
